requirements.txt
django==1.10.2
psycopg2==2.6.2
dj-database-url==0.4.1
djangorestframework==3.4.6
requests==2.11.1
